update snapshot hero
