+++ apps/web/src/pages/Pools/PoolLayout.tsx
@@ -1,21 +1,18 @@
 import React from "react";
 import { useIntl } from "react-intl";
 import { useParams, Outlet } from "react-router-dom";
-import ClipboardDocumentIcon from "@heroicons/react/24/outline/ClipboardDocumentIcon";
-import Cog8ToothIcon from "@heroicons/react/24/outline/Cog8ToothIcon";
-import UserGroupIcon from "@heroicons/react/24/outline/UserGroupIcon";
 
 import { Pending, ThrowNotFound } from "@gc-digital-talent/ui";
 
 import PageHeader from "~/components/PageHeader";
 import SEO from "~/components/SEO/SEO";
-import useRoutes from "~/hooks/useRoutes";
 import useCurrentPage from "~/hooks/useCurrentPage";
 import { Pool, useGetBasicPoolInfoQuery } from "~/api/generated";
-import { getFullPoolAdvertisementTitleLabel } from "~/utils/poolUtils";
-import { PageNavInfo } from "~/types/pages";
-
-type PageNavKeys = "view" | "edit" | "candidates";
+import {
+  getFullPoolAdvertisementTitleLabel,
+  useAdminPoolPages,
+} from "~/utils/poolUtils";
+import { PageNavKeys } from "~/types/pool";
 
 interface PoolHeaderProps {
   pool: Pick<Pool, "id" | "classifications" | "stream" | "name">;
@@ -23,57 +20,8 @@ interface PoolHeaderProps {
 
 const PoolHeader = ({ pool }: PoolHeaderProps) => {
   const intl = useIntl();
-  const paths = useRoutes();
 
-  const pages = new Map<PageNavKeys, PageNavInfo>([
-    [
-      "view",
-      {
-        icon: ClipboardDocumentIcon,
-        title: intl.formatMessage({
-          defaultMessage: "Pool information",
-          id: "Cjp2F6",
-          description: "Title for the pool info page",
-        }),
-        link: {
-          url: paths.poolView(pool.id),
-        },
-      },
-    ],
-    [
-      "edit",
-      {
-        icon: Cog8ToothIcon,
-        title: intl.formatMessage({
-          defaultMessage: "Edit pool",
-          id: "l7Wu86",
-          description: "Title for the edit pool page",
-        }),
-        link: {
-          url: paths.poolUpdate(pool.id),
-        },
-      },
-    ],
-    [
-      "candidates",
-      {
-        icon: UserGroupIcon,
-        title: intl.formatMessage({
-          defaultMessage: "Candidates",
-          id: "X4TOhW",
-          description: "Page title for the admin pool candidates index page",
-        }),
-        link: {
-          url: paths.poolCandidateTable(pool.id),
-          label: intl.formatMessage({
-            defaultMessage: "View Candidates",
-            id: "Rl+0Er",
-            description: "Title for the edit pool page",
-          }),
-        },
-      },
-    ],
-  ]);
+  const pages = useAdminPoolPages(intl, pool);
 
   const poolTitle = getFullPoolAdvertisementTitleLabel(intl, pool);
   const currentPage = useCurrentPage<PageNavKeys>(pages);
